gcb: Fixed operand generation for pseudo and compiler instructions
