+++ vadl/main/vadl/gcb/passes/operands/GenerateInstructionOperandsPass.java
@@ -18,13 +18,16 @@ package vadl.gcb.passes.operands;
 
 import static vadl.viam.ViamError.ensurePresent;
 
+import com.google.common.collect.Streams;
 import java.io.IOException;
+import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 import javax.annotation.Nullable;
 import vadl.configuration.GeneralConfiguration;
+import vadl.error.DeferredDiagnosticStore;
 import vadl.error.Diagnostic;
 import vadl.gcb.passes.operands.model.GcbInstructionImmediateOperand;
 import vadl.gcb.passes.operands.model.TableGenConstantOperand;
@@ -36,11 +39,16 @@ import vadl.gcb.passes.pseudo.PseudoFuncParamNode;
 import vadl.pass.Pass;
 import vadl.pass.PassName;
 import vadl.pass.PassResults;
+import vadl.utils.Pair;
+import vadl.viam.CompilerInstruction;
 import vadl.viam.PseudoInstruction;
 import vadl.viam.Specification;
 import vadl.viam.graph.Graph;
+import vadl.viam.graph.HasRegisterTensor;
 import vadl.viam.graph.Node;
+import vadl.viam.graph.control.InstrCallNode;
 import vadl.viam.graph.dependency.ConstantNode;
+import vadl.viam.graph.dependency.ExpressionNode;
 import vadl.viam.graph.dependency.FieldAccessRefNode;
 import vadl.viam.graph.dependency.FieldRefNode;
 import vadl.viam.graph.dependency.FuncCallNode;
@@ -88,25 +96,55 @@ public class GenerateInstructionOperandsPass extends Pass {
   }
 
   private void pseudoInstructions(Specification viam) {
-    for (var instruction : viam.isa().orElseThrow().ownPseudoInstructions()) {
-      var outputOperands = getTableGenOutputOperands(instruction.behavior());
-      var inputOperands = getTableGenInputOperands(outputOperands, instruction.behavior());
-
-      var ctx = new PseudoInstructionOperandsCtx(inputOperands, outputOperands);
-      instruction.attachExtension(ctx);
+    for (var pseudoInstruction : viam.isa().orElseThrow().ownPseudoInstructions()) {
+      attachContextToCompilerInstruction(pseudoInstruction);
     }
   }
 
   private void compilerInstructions(Specification viam) {
     var abi = viam.abi().orElseThrow();
-    for (var instruction : Stream.concat(abi.constantSequences().stream(),
+    for (var compilerInstruction : Stream.concat(abi.constantSequences().stream(),
         abi.registerAdjustmentSequences().stream()).toList()) {
-      var outputOperands = getTableGenOutputOperands(instruction.behavior());
-      var inputOperands = getTableGenInputOperands(outputOperands, instruction.behavior());
-
-      var ctx = new PseudoInstructionOperandsCtx(inputOperands, outputOperands);
-      instruction.attachExtension(ctx);
+      attachContextToCompilerInstruction(compilerInstruction);
     }
+  }
+
+  private void attachContextToCompilerInstruction(CompilerInstruction compilerInstruction) {
+    var totalInputs = new ArrayList<TableGenInstructionOperand>();
+    var totalOutputs = new ArrayList<TableGenInstructionOperand>();
+
+    for (var callNode : compilerInstruction.behavior().getNodes(InstrCallNode.class).toList()) {
+      var instruction = callNode.target();
+      var instructionBehavior = instruction.behavior().copy();
+      replaceNodesInBehavior(instructionBehavior, callNode);
+
+      var outputOperands = getTableGenOutputOperands(instructionBehavior);
+      var inputOperands = getTableGenInputOperands(outputOperands, instructionBehavior);
+
+      addWithoutDuplicates(totalInputs, inputOperands);
+      addWithoutDuplicates(totalOutputs, outputOperands);
+    }
+
+    var ctx = new CompilerInstructionOperandsCtx(totalInputs, totalOutputs);
+    compilerInstruction.attachExtension(ctx);
+  }
+
+  private void addWithoutDuplicates(List<TableGenInstructionOperand> dest,
+                                    List<TableGenInstructionOperand> src) {
+    for (var element : src) {
+      boolean exists = false;
+      for (var needle : dest) {
+        if (needle.equals(element)) {
+          exists = true;
+          break;
+        }
+      }
+
+      if (!exists) {
+        dest.add(element);
+      }
+    }
+
   }
 
   private List<TableGenInstructionOperand> getTableGenOutputOperands(Graph behavior) {
@@ -288,5 +326,109 @@ public class GenerateInstructionOperandsPass extends Pass {
     return graph.getNodes(WriteRegTensorNode.class)
         .filter(e -> e.regTensor().isRegisterFile())
         .toList();
+  }
+
+  /**
+   * There are two relevant cases.
+   * The first is that the {@code argument} is a constant. Then, we do not have to do anything.
+   * The second case is when {@link CompilerInstruction} uses an {@code index}. Then, the argument
+   * is replaced by a {@link FuncParamNode}. However, we still require to know the index for
+   * the pseudo instance expansion. That's why we extend {@link FuncParamNode} with
+   * {@link PseudoFuncParamNode} which has an {@code index} property.
+   * Here is an example of the index. Note that {@code rs} will be transformed into
+   * a {@link PseudoFuncParamNode} when it is replaced.
+   * <code>
+   * pseudo instruction BGEZ( rs : Index, offset : Bits<12> ) =
+   * {
+   * BGE{ rs1 = rs, rs2 = 0 as Bits5, imm = offset }
+   * }
+   * </code>
+   */
+  protected static ExpressionNode indexArgument(List<ExpressionNode> arguments,
+                                                ExpressionNode argument) {
+    if (argument instanceof FuncParamNode funcParamNode) {
+      int index = arguments.indexOf(argument);
+      return new PseudoFuncParamNode(funcParamNode.parameter(), index);
+    }
+    return argument;
+  }
+
+  /**
+   * Replace the arguments in the behavior of {@code copiedInstructionBehavior}.
+   */
+  public static void replaceNodesInBehavior(Graph copiedInstructionBehavior,
+                                            InstrCallNode callNode) {
+    Streams.zip(callNode.getParamFields().stream(), callNode.arguments().stream(),
+            Pair::new)
+        .forEach(app -> {
+          var formatField = app.left();
+          var argument = indexArgument(callNode.arguments(), app.right());
+
+          Stream.concat(
+                  copiedInstructionBehavior.getNodes(FieldRefNode.class),
+                  copiedInstructionBehavior.getNodes(FieldAccessRefNode.class)
+              )
+              .filter(x -> {
+                if (x instanceof FieldRefNode fieldRefNode) {
+                  return fieldRefNode.formatField().equals(formatField);
+                } else if (x instanceof FieldAccessRefNode fieldAccessRefNode) {
+                  return fieldAccessRefNode.fieldAccess().fieldRefs().contains(formatField);
+                }
+                return false;
+              })
+              .forEach(occurrence -> {
+                // Edge case:
+                // When we have the following pseudo instruction. Note that "r1" is replaced
+                // by a constant. Sometimes, we need to create instruction selectors in TableGen,
+                // and it requires a variable. However, if we replace the field by a constant
+                // we lose the name of the variable because we have no field anymore.
+                // {
+                //     JALR{ rs1 = 1 as Bits5, rd = 0 as Bits5, imm = 0 as Bits12 }
+                // }
+
+                if (argument instanceof ConstantNode constantNode) {
+                  // The constantNode tells me that it will be used as a register index.
+
+                  // Go over the usages to emit warnings.
+                  // We need the usage because we need to find out what the register file
+                  // to check for constraints.
+                  occurrence.usages()
+                      .filter(node -> (node instanceof HasRegisterTensor x && x.hasRegisterFile()))
+                      .forEach(node -> {
+                        var cast = (HasRegisterTensor) node;
+
+                        var constraintValue =
+                            Arrays.stream(cast.registerTensor().constraints()).filter(
+                                c -> c.indices().getFirst().intValue()
+                                    == constantNode.constant().asVal().intValue()).findFirst();
+
+                        if (constraintValue.isEmpty()) {
+                          DeferredDiagnosticStore.add(Diagnostic.warning(
+                              "There is no constraint value for this register. "
+                                  +
+                                  "Therefore, we cannot generate instruction selectors for it.",
+                              occurrence.location()).build());
+                        }
+                      });
+
+                  occurrence.replaceAndDelete(argument.copy());
+
+                  // After the replacement, we can check whether we have a write node with
+                  // constant node as address which has a constraint. If that's the case, then we
+                  // can remove the side effect.
+                  occurrence.usages()
+                      .filter(node -> node instanceof WriteRegTensorNode writeRegTensorNode
+                          && writeRegTensorNode.regTensor().isRegisterFile()
+                          && writeRegTensorNode.hasConstantAddress()
+                          // Check if there is a constraint for this register index.
+                          && Arrays.stream(writeRegTensorNode.regTensor().constraints())
+                          .anyMatch(constraint -> constraint.indices().getFirst().intValue()
+                              == constantNode.constant().asVal().intValue()))
+                      .forEach(Node::safeDelete);
+                } else {
+                  occurrence.replaceAndDelete(argument.copy());
+                }
+              });
+        });
   }
 }
