+++ vadl/main/vadl/lcb/passes/llvmLowering/strategies/LlvmCompilerInstructionLowerStrategy.java
@@ -16,37 +16,22 @@
 
 package vadl.lcb.passes.llvmLowering.strategies;
 
-import com.google.common.collect.Streams;
 import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.LinkedHashSet;
 import java.util.List;
 import java.util.Optional;
-import java.util.stream.Stream;
 import vadl.error.DeferredDiagnosticStore;
 import vadl.error.Diagnostic;
 import vadl.gcb.passes.DetermineRegisterUsesAndDefsPass;
 import vadl.gcb.passes.RegisterRef;
-import vadl.gcb.passes.operands.model.TableGenDefaultInstructionOperand;
-import vadl.gcb.passes.operands.model.TableGenInstructionImmediateOperand;
-import vadl.gcb.passes.operands.model.TableGenInstructionOperand;
-import vadl.gcb.passes.pseudo.PseudoFuncParamNode;
+import vadl.gcb.passes.operands.CompilerInstructionOperandsCtx;
+import vadl.gcb.passes.operands.GenerateInstructionOperandsPass;
 import vadl.lcb.passes.isaMatching.IsaMachineInstructionMatchingPass;
 import vadl.lcb.passes.llvmLowering.LlvmLoweringPass;
 import vadl.lcb.passes.llvmLowering.domain.LlvmLoweringRecord;
-import vadl.utils.Pair;
 import vadl.viam.CompilerInstruction;
-import vadl.viam.Parameter;
-import vadl.viam.graph.Graph;
-import vadl.viam.graph.HasRegisterTensor;
-import vadl.viam.graph.Node;
+import vadl.viam.ViamError;
 import vadl.viam.graph.control.InstrCallNode;
-import vadl.viam.graph.dependency.ConstantNode;
-import vadl.viam.graph.dependency.ExpressionNode;
-import vadl.viam.graph.dependency.FieldAccessRefNode;
-import vadl.viam.graph.dependency.FieldRefNode;
-import vadl.viam.graph.dependency.FuncParamNode;
-import vadl.viam.graph.dependency.WriteRegTensorNode;
 
 /**
  * Abstract class to lower {@link CompilerInstruction} to {@link LlvmLoweringRecord.Compiler}.
@@ -69,8 +54,6 @@ public abstract class LlvmCompilerInstructionLowerStrategy {
   ) {
     var uses = new ArrayList<RegisterRef>();
     var defs = new ArrayList<RegisterRef>();
-    var inputOperands = new ArrayList<TableGenInstructionOperand>();
-    var outputOperands = new ArrayList<TableGenInstructionOperand>();
 
     var isTerminator = false;
     var isReturn = false;
@@ -90,8 +73,7 @@ public abstract class LlvmCompilerInstructionLowerStrategy {
     for (var callNode : compilerInstruction.behavior().getNodes(InstrCallNode.class).toList()) {
       var instruction = callNode.target();
       var instructionBehavior = instruction.behavior().copy();
-
-      replaceNodesInBehavior(instructionBehavior, callNode);
+      GenerateInstructionOperandsPass.replaceNodesInBehavior(instructionBehavior, callNode);
 
       var label = supportedInstructions.reverse().get(instruction);
 
@@ -117,37 +99,8 @@ public abstract class LlvmCompilerInstructionLowerStrategy {
         defs.addAll(baseInstructionInfo.defs());
         uses.addAll(baseInstructionInfo.uses());
 
-        addWithoutDuplicates(inputOperands, baseInstructionInfo.inputs());
-        addWithoutDuplicates(outputOperands, baseInstructionInfo.outputs());
-
         break;
       }
-    }
-
-    // We have now computed both `inputOperands` and `outputOperands` based on the
-    // underlying machine instructions. However, those aren't the real operands of the
-    // pseudo instruction. The pseudo instruction defines its own operands.
-
-    /*
-      Pseudo Instruction Operands (out): rd ; (in): symbol
-      Machine Instructions (out): rd ; (in): rs1, imm
-
-      pseudo instruction LGA_64( rd: Index, symbol: Bits<32> ) =
-      {
-          AUIPC { rd = rd, imm = got_pcrel_hi( symbol ) }
-          LD { rd = rd, rs1 = rd, imm = pcrel_lo( symbol ) }
-      }
-      assembly LGA_64 = ("LGA", " ", register(rd), ",", hex( symbol ))
-     */
-
-    var compilerInstructionInputOperands = new ArrayList<TableGenInstructionOperand>();
-    var compilerInstructionsOutputOperands = new ArrayList<TableGenInstructionOperand>();
-
-    for (var parameter : compilerInstruction.parameters()) {
-      matchesParameter(parameter, inputOperands)
-          .ifPresent(compilerInstructionInputOperands::add);
-      matchesParameter(parameter, outputOperands)
-          .ifPresent(compilerInstructionsOutputOperands::add);
     }
 
     var flags = new LlvmLoweringPass.Flags(isTerminator,
@@ -162,9 +115,14 @@ public abstract class LlvmCompilerInstructionLowerStrategy {
         false,
         false);
 
+    var operandCtx =
+        ViamError.ensureNonNull(compilerInstruction.extension(CompilerInstructionOperandsCtx.class),
+            () -> Diagnostic.error("Cannot lookup operands for instruction",
+                compilerInstruction.location()));
+
     var info = new LlvmLoweringPass.BaseInstructionInfo(
-        dedup(compilerInstructionInputOperands),
-        dedup(compilerInstructionsOutputOperands),
+        operandCtx.inputs(),
+        operandCtx.outputs(),
         flags,
         dedup(uses),
         dedup(defs)
@@ -173,142 +131,6 @@ public abstract class LlvmCompilerInstructionLowerStrategy {
     return Optional.of(new LlvmLoweringRecord.Compiler(
         info
     ));
-  }
-
-  private Optional<TableGenInstructionOperand> matchesParameter(Parameter parameter,
-                                                                List<TableGenInstructionOperand> operands) {
-    for (var operand : operands) {
-      if (operand instanceof TableGenDefaultInstructionOperand defaultInstructionOperand
-          && defaultInstructionOperand.name().equals(parameter.simpleName())) {
-        return Optional.of(operand);
-      } else if (operand instanceof TableGenInstructionImmediateOperand immediateOperand
-          && immediateOperand.name().equals(parameter.simpleName())) {
-        return Optional.of(operand);
-      }
-    }
-
-    return Optional.empty();
-  }
-
-  private void addWithoutDuplicates(List<TableGenInstructionOperand> dest,
-                                    List<TableGenInstructionOperand> src) {
-    for (var element : src) {
-      boolean exists = false;
-      for (var needle : dest) {
-        if (needle.equals(element)) {
-          exists = true;
-          break;
-        }
-      }
-
-      if (!exists) {
-        dest.add(element);
-      }
-    }
-
-  }
-
-  /**
-   * There are two relevant cases.
-   * The first is that the {@code argument} is a constant. Then, we do not have to do anything.
-   * The second case is when {@link CompilerInstruction} uses an {@code index}. Then, the argument
-   * is replaced by a {@link FuncParamNode}. However, we still require to know the index for
-   * the pseudo instance expansion. That's why we extend {@link FuncParamNode} with
-   * {@link PseudoFuncParamNode} which has an {@code index} property.
-   * Here is an example of the index. Note that {@code rs} will be transformed into
-   * a {@link PseudoFuncParamNode} when it is replaced.
-   * <code>
-   * pseudo instruction BGEZ( rs : Index, offset : Bits<12> ) =
-   * {
-   * BGE{ rs1 = rs, rs2 = 0 as Bits5, imm = offset }
-   * }
-   * </code>
-   */
-  protected ExpressionNode indexArgument(List<ExpressionNode> arguments, ExpressionNode argument) {
-    if (argument instanceof FuncParamNode funcParamNode) {
-      int index = arguments.indexOf(argument);
-      return new PseudoFuncParamNode(funcParamNode.parameter(), index);
-    }
-    return argument;
-  }
-
-  /**
-   * Replace the arguments in the behavior of {@code copiedInstructionBehavior}.
-   */
-  public void replaceNodesInBehavior(Graph copiedInstructionBehavior,
-                                     InstrCallNode callNode) {
-    Streams.zip(callNode.getParamFields().stream(), callNode.arguments().stream(),
-            Pair::new)
-        .forEach(app -> {
-          var formatField = app.left();
-          var argument = indexArgument(callNode.arguments(), app.right());
-
-          Stream.concat(
-                  copiedInstructionBehavior.getNodes(FieldRefNode.class),
-                  copiedInstructionBehavior.getNodes(FieldAccessRefNode.class)
-              )
-              .filter(x -> {
-                if (x instanceof FieldRefNode fieldRefNode) {
-                  return fieldRefNode.formatField().equals(formatField);
-                } else if (x instanceof FieldAccessRefNode fieldAccessRefNode) {
-                  return fieldAccessRefNode.fieldAccess().fieldRefs().contains(formatField);
-                }
-                return false;
-              })
-              .forEach(occurrence -> {
-                // Edge case:
-                // When we have the following pseudo instruction. Note that "r1" is replaced
-                // by a constant. Sometimes, we need to create instruction selectors in TableGen,
-                // and it requires a variable. However, if we replace the field by a constant
-                // we lose the name of the variable because we have no field anymore.
-                // {
-                //     JALR{ rs1 = 1 as Bits5, rd = 0 as Bits5, imm = 0 as Bits12 }
-                // }
-
-                if (argument instanceof ConstantNode constantNode) {
-                  // The constantNode tells me that it will be used as a register index.
-
-                  // Go over the usages to emit warnings.
-                  // We need the usage because we need to find out what the register file
-                  // to check for constraints.
-                  occurrence.usages()
-                      .filter(node -> (node instanceof HasRegisterTensor x && x.hasRegisterFile()))
-                      .forEach(node -> {
-                        var cast = (HasRegisterTensor) node;
-
-                        var constraintValue =
-                            Arrays.stream(cast.registerTensor().constraints()).filter(
-                                c -> c.indices().getFirst().intValue()
-                                    == constantNode.constant().asVal().intValue()).findFirst();
-
-                        if (constraintValue.isEmpty()) {
-                          DeferredDiagnosticStore.add(Diagnostic.warning(
-                              "There is no constraint value for this register. "
-                                  +
-                                  "Therefore, we cannot generate instruction selectors for it.",
-                              occurrence.location()).build());
-                        }
-                      });
-
-                  occurrence.replaceAndDelete(argument.copy());
-
-                  // After the replacement, we can check whether we have a write node with
-                  // constant node as address which has a constraint. If that's the case, then we
-                  // can remove the side effect.
-                  occurrence.usages()
-                      .filter(node -> node instanceof WriteRegTensorNode writeRegTensorNode
-                          && writeRegTensorNode.regTensor().isRegisterFile()
-                          && writeRegTensorNode.hasConstantAddress()
-                          // Check if there is a constraint for this register index.
-                          && Arrays.stream(writeRegTensorNode.regTensor().constraints())
-                          .anyMatch(constraint -> constraint.indices().getFirst().intValue()
-                              == constantNode.constant().asVal().intValue()))
-                      .forEach(Node::safeDelete);
-                } else {
-                  occurrence.replaceAndDelete(argument.copy());
-                }
-              });
-        });
   }
 
   private <T> List<T> dedup(
